+++ src/test/java/com/jprotractor/integration/NgCalculatorTest.java
@@ -1,0 +1,142 @@
+package com.jprotractor.integration;
+
+import org.apache.commons.lang.exception.ExceptionUtils;
+import static org.hamcrest.CoreMatchers.containsString;
+import static org.hamcrest.CoreMatchers.equalTo;
+import static org.hamcrest.CoreMatchers.notNullValue;
+
+import static org.junit.Assert.assertNotEquals;
+import static org.junit.Assert.assertThat;
+import static org.junit.Assert.assertTrue;
+
+import java.io.IOException;
+
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.Enumeration;
+import java.util.Formatter;
+import java.util.Iterator;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+
+import java.util.concurrent.TimeUnit;
+
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
+import org.junit.After;
+import org.junit.AfterClass;
+import org.junit.Before;
+import org.junit.BeforeClass;
+import org.junit.experimental.categories.Category;
+import org.junit.experimental.runners.Enclosed;
+import org.junit.runner.RunWith;
+import org.junit.Test;
+
+import org.openqa.selenium.Alert;
+import org.openqa.selenium.By;
+import org.openqa.selenium.Dimension;
+import org.openqa.selenium.firefox.FirefoxProfile;
+import org.openqa.selenium.firefox.internal.ProfilesIni;
+import org.openqa.selenium.interactions.Actions;
+import org.openqa.selenium.JavascriptExecutor;
+import org.openqa.selenium.NoAlertPresentException;
+import org.openqa.selenium.NoSuchElementException;
+import org.openqa.selenium.phantomjs.PhantomJSDriver;
+import org.openqa.selenium.Platform;
+import org.openqa.selenium.remote.DesiredCapabilities;
+import org.openqa.selenium.remote.RemoteWebDriver;
+import org.openqa.selenium.SearchContext;
+import org.openqa.selenium.support.ui.ExpectedConditions;
+import org.openqa.selenium.support.ui.Select;
+import org.openqa.selenium.support.ui.WebDriverWait;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.WebDriverException;
+import org.openqa.selenium.WebElement;
+
+import com.jprotractor.NgBy;
+import com.jprotractor.NgWebDriver;
+import com.jprotractor.NgWebElement;
+
+public class NgCalculatorTest {
+	private static String fullStackTrace;
+	private static NgWebDriver ngDriver;
+	private static WebDriver seleniumDriver;
+	static WebDriverWait wait;
+	static Actions actions;
+	static Alert alert;
+	static int implicitWait = 10;
+	static int flexibleWait = 5;
+	static long pollingInterval = 500;
+	static int width = 600;
+	static int height = 400;
+	// set to true for Desktop, false for headless browser testing
+	static boolean isCIBuild =  false;
+    private static String baseUrl = "http://juliemr.github.io/protractor-demo/";
+
+	@BeforeClass
+	public static void setup() throws IOException{
+		isCIBuild = CommonFunctions.checkEnvironment();
+		seleniumDriver = CommonFunctions.getSeleniumDriver();
+		seleniumDriver.manage().window().setSize(new Dimension(width , height ));
+		seleniumDriver.manage().timeouts().pageLoadTimeout(50, TimeUnit.SECONDS).implicitlyWait(implicitWait, TimeUnit.SECONDS).setScriptTimeout(10, TimeUnit.SECONDS);
+		wait = new WebDriverWait(seleniumDriver, flexibleWait );
+		wait.pollingEvery(pollingInterval,TimeUnit.MILLISECONDS);
+		actions = new Actions(seleniumDriver);		
+		ngDriver = new NgWebDriver(seleniumDriver);
+	}
+
+	@Before
+	public void beforeEach() {
+		ngDriver.navigate().to(baseUrl);
+	}
+
+	@Test
+	public void testAddition() throws Exception {
+		NgWebElement element = ngDriver.findElement(NgBy.model("first"));
+		assertThat(element,notNullValue());
+		highlight(element, 100);
+		element.sendKeys("40");
+		element = ngDriver.findElement(NgBy.model("second"));
+		assertThat(element,notNullValue());
+		highlight(element, 100);
+		element.sendKeys("2");
+		element = ngDriver.findElement(NgBy.options("value for (key, value) in operators"));
+		assertThat(element,notNullValue());
+		element.click();
+		element = ngDriver.findElement(NgBy.buttonText("Go!"));
+		assertThat(element,notNullValue());
+		assertThat(element.getText(),containsString("Go"));
+		element.click();
+		element = ngDriver.findElement(NgBy.binding("latest" ));
+		assertThat(element, notNullValue());
+		assertThat(element.getText(), equalTo("42"));
+		highlight(element, 100);
+	}
+
+	@AfterClass
+	public static void teardown() {
+		ngDriver.close();
+		seleniumDriver.quit();		
+	}
+
+	private static void highlight(WebElement element) throws InterruptedException {
+		highlight(element,  100);
+	}
+
+	private static void highlight(WebElement element, long highlightInterval ) throws InterruptedException {
+		CommonFunctions.highlight(element, highlightInterval);
+	}
+
+	private static void highlight(NgWebElement element) throws InterruptedException {
+		highlight(element,  100);
+	}
+
+	private static void highlight(NgWebElement element, long highlightInterval ) throws InterruptedException {
+		CommonFunctions.highlight(element, highlightInterval);
+	}
+	
+}
+
+
